pyproject.toml
[tool.poetry]
name = "livinebackend"
version = "0.1.0"
description = ""
authors = ["Mazen Omar <wildlifemain1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.2"
djangorestframework = "^3.14.0"
django-rest-knox = "^4.2.0"
pillow = "^9.3.0"
django-rest-passwordreset = "^1.3.0"
whitenoise = "^6.2.0"
python-decouple = "^3.6"
six = "^1.16.0"
django-modeltranslation = "^0.18.9"
mysqlclient = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
